src/main/java/de/sebastianmonzel/textextractor/TxtTextExtractor.java
package de.sebastianmonzel.textextractor;

import java.io.*;
import java.nio.file.Files;
import java.nio.file.Path;


public class TxtTextExtractor extends TextExtractor {

    @Override
    public String extractText(File file) {
        try {
            Path path = file.toPath();
            byte[] bytes = Files.readAllBytes(path);
            String content = new String(bytes);

            return content;
        } catch (IOException e) {
            e.printStackTrace();
        }
        return null;
    }

    @Override
    public String extractText(InputStream inputStream) {
        BufferedReader br = new BufferedReader(new InputStreamReader(inputStream));
        StringBuilder stringBuilder = new StringBuilder();
        try {
            String line;
            while ((line = br.readLine()) != null) {
                stringBuilder.append(line);
            }
        } catch (IOException e) {
            e.printStackTrace();
        } finally {
            try {
                br.close();
            } catch (IOException e) {
                e.printStackTrace();
            }
        }
        return stringBuilder.toString();
    }


}
